Add support to post github webhook events in chat
Webhook format: `https://your-bot-domain.com/api/webhook/github/{channel_id}`

Secret is configured in the config under Auth -> GitHub -> Webhook -> Secret

update example config
